Added support for DateTime and Date objects.
`ActiveSupport` adds `#utc` to `DateTime` and `Date` that allow both of
these objects to work with this gem instead of raising errors. To avoid
confusion and provide support for `DateTime` and `Date` in regular Ruby,
both of these objects can now be provided instead of a `Time` object. To
allow for forwards and backwards compatibility the object is always
converted into a time object.

Closes #36