Add to_s and inspect to Timezone::Zone
The current inspect, in particular, is unwieldy because it prints out
all of the rules. I've implemented a simple `to_s` that prints the zone
name and a nicer `inspect` that simply prints `rules: [...]` instead of
listing all the rules. From that point a debugger can look further into
the rules if they so chose.