[0.99] Deprecate Zone#active_support_time_zone
This whole collection of functionality is too tightly coupled to Rails.
It is also unmaintainable due to possible forwards-backwards
incompatible changes on the `ActiveSupport` side (basically we don't
know what version of Rails someone might be using so we can't keep the
hash up to date).

For some more background, see #43.