[0.99] Deprecate Zone::list
This method is overly complex for a very small benefit. The entire hash
structure, sorting and replacements can all easily be implemented by a
consumer of this gem using publicly available methods. This provides a
consumer with much finer control over which information to include, the
API to use, and whether or not they even want to use replacements. The
code is also currently not very secure and supporting it in a stable
release is going to be a headache.