Improve Geonames exception messaging
Geonames is pretty good about returning error messages. #54 clearly
shows that exception messages are available, so we should be exposing
these to the calling code (albeit wrapped in a
`Timezone::Error::GeoNames` exception).

Previously, all but the API limit exceptions were swallowed up and
`#lookup` returned `nil` which resulted in an `InvalidZone` error.
Oftentimes the zone was not invalid, and in fact there was a real error
message from geonames.

This change modifies the lookup flow so that if a body is present, it is
immediately parsed and we try to extract a `timezoneId`. If one is not
present, then we search the parsed JSON for a `status` and
`status.message` field and wrap that in our exception.

The main change here is that if a response code is not an HTTP `2**`
code, we still try to parse the body. I think this is a better flow as
there are cases where geonames returns a `503` HTTP code with a valid
error message. Additionally, wrapping any exception messages from
geonames allows us to provide more details to the consumer about what
has gone wrong with their request.

Fixes #54