Handle GeoNames Error Code 15
This error code is returned when the lat/lng resolves to an actual point
on the planet which has no associated timezone information. It is more
appropriate in this case to return `nil` (which will result in an
`InvalidZone` error) instead of raising a `GeoNames` exception because
the lat/lng is valid, the lookup has succeeded, but there is no
associated timezone information for the point.