Remove Cached Rules from Marshal Dump
Whenever a calculation on `Zone` requires access to timezone rules, it
fetches them from `Loader::load` and caches them in `@rules` using the
`private_rules` method. Unfortunately, this cached data is retained
in the `Marshal::dump` output for a `Zone` even though it is entirely
unnecessary. To avoid any backwards incompatibility, while still
removing the cached `@rules` from the dump, always lookup the rule name
from the `Loader`. In order to retain performance I have also done a
little re-working of the `Loader::load` method so that it only does a
single `Hash` lookup after the initial cache population.

Fixes #97